+++ libs/data-mapper-v2/src/core/state/DataMapSlice.ts
@@ -13,7 +13,12 @@ import type { UnknownNode } from '../../utils/DataMap.Utils';
 import { getParentId } from '../../utils/DataMap.Utils';
 import { getConnectedSourceSchema, getFunctionLocationsForAllFunctions, isFunctionData } from '../../utils/Function.Utils';
 import { LogService } from '../../utils/Logging.Utils';
-import { flattenSchemaIntoDictionary, flattenSchemaIntoSortArray, isSchemaNodeExtended } from '../../utils/Schema.Utils';
+import {
+  flattenSchemaIntoDictionary,
+  flattenSchemaIntoSortArray,
+  flattenSchemaNodeMap,
+  isSchemaNodeExtended,
+} from '../../utils/Schema.Utils';
 import type { FunctionMetadata, MapMetadata, SchemaExtended, SchemaNodeDictionary, SchemaNodeExtended } from '@microsoft/logic-apps-shared';
 import { SchemaNodeProperty, SchemaType } from '@microsoft/logic-apps-shared';
 import type { PayloadAction } from '@reduxjs/toolkit';
@@ -30,6 +35,7 @@ export interface DataMapState {
 
 export interface DataMapOperationState {
   dataMapConnections: ConnectionDictionary;
+  dataMapLML: string;
   sourceSchema?: SchemaExtended;
   flattenedSourceSchema: SchemaNodeDictionary;
   sourceSchemaOrdering: string[];
@@ -52,6 +58,7 @@ export interface DataMapOperationState {
 
 const emptyPristineState: DataMapOperationState = {
   dataMapConnections: {},
+  dataMapLML: '',
   currentSourceSchemaNodes: [],
   functionNodes: {},
   flattenedSourceSchema: {},
@@ -125,31 +132,49 @@ export const dataMapSlice = createSlice({
 
     setInitialSchema: (state, action: PayloadAction<InitialSchemaAction>) => {
       const flattenedSchema = flattenSchemaIntoDictionary(action.payload.schema, action.payload.schemaType);
+      const currentState = state.curDataMapOperation;
 
       if (action.payload.schemaType === SchemaType.Source) {
         const sourceSchemaSortArray = flattenSchemaIntoSortArray(action.payload.schema.schemaTreeRoot);
+        const sourceCurrentFlattenedSchemaMap = currentState.sourceSchema
+          ? flattenSchemaNodeMap(currentState.sourceSchema.schemaTreeRoot)
+          : {};
 
-        state.curDataMapOperation.sourceSchema = action.payload.schema;
-        state.curDataMapOperation.flattenedSourceSchema = flattenedSchema;
-        state.curDataMapOperation.sourceSchemaOrdering = sourceSchemaSortArray;
+        currentState.sourceSchema = action.payload.schema;
+        currentState.flattenedSourceSchema = flattenedSchema;
+        currentState.sourceSchemaOrdering = sourceSchemaSortArray;
         state.pristineDataMap.sourceSchema = action.payload.schema;
         state.pristineDataMap.flattenedSourceSchema = flattenedSchema;
         state.pristineDataMap.sourceSchemaOrdering = sourceSchemaSortArray;
+
+        // NOTE: Reset ReactFlow nodes to filter out source nodes
+        currentState.nodes = currentState.nodes.filter((node) => !sourceCurrentFlattenedSchemaMap[node.data.id]);
       } else {
         const targetSchemaSortArray = flattenSchemaIntoSortArray(action.payload.schema.schemaTreeRoot);
+        const targetCurrentFlattenedSchemaMap = currentState.targetSchema
+          ? flattenSchemaNodeMap(currentState.targetSchema.schemaTreeRoot)
+          : {};
 
-        state.curDataMapOperation.targetSchema = action.payload.schema;
-        state.curDataMapOperation.flattenedTargetSchema = flattenedSchema;
-        state.curDataMapOperation.targetSchemaOrdering = targetSchemaSortArray;
-        state.curDataMapOperation.currentTargetSchemaNode = undefined;
+        currentState.targetSchema = action.payload.schema;
+        currentState.flattenedTargetSchema = flattenedSchema;
+        currentState.targetSchemaOrdering = targetSchemaSortArray;
+        currentState.currentTargetSchemaNode = undefined;
         state.pristineDataMap.targetSchema = action.payload.schema;
         state.pristineDataMap.flattenedTargetSchema = flattenedSchema;
         state.pristineDataMap.targetSchemaOrdering = targetSchemaSortArray;
+
+        // NOTE: Reset ReactFlow nodes to filter out source nodes
+        currentState.nodes = currentState.nodes.filter((node) => !targetCurrentFlattenedSchemaMap[node.data.id]);
       }
 
+      // NOTE: Reset ReactFlow edges
+      currentState.edges = [];
+
       if (state.curDataMapOperation.sourceSchema && state.curDataMapOperation.targetSchema) {
-        state.curDataMapOperation.currentTargetSchemaNode = state.curDataMapOperation.targetSchema.schemaTreeRoot;
+        currentState.currentTargetSchemaNode = state.curDataMapOperation.targetSchema.schemaTreeRoot;
       }
+
+      state.curDataMapOperation = { ...currentState };
     },
 
     setInitialDataMap: (state, action: PayloadAction<InitialDataMapAction>) => {
@@ -195,14 +220,14 @@ export const dataMapSlice = createSlice({
       if (action.payload.reactFlowSource.startsWith(SchemaType.Source)) {
         sourceNode = state.curDataMapOperation.flattenedSourceSchema[action.payload.reactFlowSource];
       } else {
-        sourceNode = newState.functionNodes[action.payload.reactFlowSource].functionData;
+        sourceNode = newState.functionNodes[action.payload.reactFlowSource]?.functionData;
       }
       let destinationNode: UnknownNode;
 
       if (action.payload.reactFlowDestination.startsWith(SchemaType.Target)) {
         destinationNode = state.curDataMapOperation.flattenedTargetSchema[action.payload.reactFlowDestination];
       } else {
-        destinationNode = newState.functionNodes[action.payload.reactFlowSource].functionData;
+        destinationNode = newState.functionNodes[action.payload.reactFlowSource]?.functionData;
       }
 
       addConnection(newState.dataMapConnections, action.payload, destinationNode, sourceNode);
@@ -217,9 +242,16 @@ export const dataMapSlice = createSlice({
       doDataMapOperation(state, newState, 'Make connection');
     },
 
+    updateDataMapLML: (state, action: PayloadAction<string>) => {
+      state.curDataMapOperation.dataMapLML = action.payload;
+    },
+
     saveDataMap: (
       state,
-      action: PayloadAction<{ sourceSchemaExtended: SchemaExtended | undefined; targetSchemaExtended: SchemaExtended | undefined }>
+      action: PayloadAction<{
+        sourceSchemaExtended: SchemaExtended | undefined;
+        targetSchemaExtended: SchemaExtended | undefined;
+      }>
     ) => {
       const sourceSchemaExtended = action.payload.sourceSchemaExtended;
       const targetSchemaExtended = action.payload.targetSchemaExtended;