+++ libs/designer-ui/src/lib/templates/index.ts
@@ -1,4 +1,4 @@
-export * from './templatePanelUtil';
+export * from './model';
 export * from './templatesPanelContent';
 export * from './templatesParametersField';
 export * from './templatesPanelFooter';