+++ libs/designer/src/lib/core/state/templates/templateSlice.ts
@@ -5,6 +5,7 @@ import {
   InitConnectionService,
   InitFunctionService,
   InitGatewayService,
+  InitTenantService,
   InitOAuthService,
   getIntl,
   getRecordEntry,
@@ -28,6 +29,7 @@ interface TemplateData {
     validationErrors: Record<string, string | undefined>;
   };
   connections: Record<string, Template.Connection>;
+  images?: Record<string, any>;
 }
 
 export interface TemplateState extends TemplateData {
@@ -46,6 +48,7 @@ const initialState: TemplateState = {
   },
   connections: {},
   servicesInitialized: false,
+  images: {},
 };
 
 export const initializeTemplateServices = createAsyncThunk(
@@ -54,6 +57,7 @@ export const initializeTemplateServices = createAsyncThunk(
     connectionService,
     oAuthService,
     gatewayService,
+    tenantService,
     apimService,
     functionService,
     appServiceService,
@@ -64,6 +68,9 @@ export const initializeTemplateServices = createAsyncThunk(
 
     if (gatewayService) {
       InitGatewayService(gatewayService);
+    }
+    if (tenantService) {
+      InitTenantService(tenantService);
     }
     if (apimService) {
       InitApiManagementService(apimService);
@@ -172,6 +179,7 @@ export const templateSlice = createSlice({
         state.manifest = action.payload.manifest;
         state.parameters = action.payload.parameters;
         state.connections = action.payload.connections;
+        state.images = action.payload.images;
       }
     });
 
@@ -210,6 +218,12 @@ const loadTemplateFromGithub = async (templateName: string, manifest: Template.M
 
     const templateManifest: Template.Manifest =
       manifest ?? (await import(`${templatesPathFromState}/${templateName}/manifest.json`)).default;
+
+    const images: Record<string, any> = {};
+    for (const key of Object.keys(templateManifest.images)) {
+      images[key] = (await import(`${templatesPathFromState}/${templateName}/${templateManifest.images[key]}`)).default;
+    }
+
     const parametersDefinitions = templateManifest.parameters?.reduce((result: Record<string, Template.ParameterDefinition>, parameter) => {
       result[parameter.name] = {
         ...parameter,
@@ -228,6 +242,7 @@ const loadTemplateFromGithub = async (templateName: string, manifest: Template.M
         validationErrors: {},
       },
       connections: templateManifest.connections,
+      images,
     };
   } catch (ex) {
     console.error(ex);