+++ libs/logic-apps-shared/src/designer-client-services/index.ts
@@ -24,3 +24,4 @@ export * from './lib/editor';
 export * from './lib/connectionParameterEditor';
 export * from './lib/chatbot';
 export * from './lib/customcode';
+export * from './lib/tenant';