+++ CHANGELOG.md
@@ -1,4 +1,23 @@
 # Logic Apps Designer
+## [4.32.0](https://github.com/Azure/LogicAppsUX/compare/v4.31.0...v4.32.0) (2024-06-20)
+
+
+### Features
+
+* **Data Mapper:** Handle Node connections and state management ([#4990](https://github.com/Azure/LogicAppsUX/issues/4990)) ([dd0c2b8](https://github.com/Azure/LogicAppsUX/commit/dd0c2b86674068a94eebe1a011b50d9001ee764f))
+* **designer-ui:** Allow makers to use `class` attributes and `table` styling in raw HTML ([#5000](https://github.com/Azure/LogicAppsUX/issues/5000)) ([3446119](https://github.com/Azure/LogicAppsUX/commit/344611928460671e46ae740923c0869d2732a606))
+* **template:** Updating quickview panel with latest UI ([#4994](https://github.com/Azure/LogicAppsUX/issues/4994)) ([4704eaf](https://github.com/Azure/LogicAppsUX/commit/4704eafc0dcce8c7026beba7cfe0e6715a49a25a))
+
+## [4.31.0](https://github.com/Azure/LogicAppsUX/compare/v4.30.0...v4.31.0) (2024-06-19)
+
+
+### Features
+
+* **Data Mapper:** basic version of code view ([#4996](https://github.com/Azure/LogicAppsUX/issues/4996)) ([ce2ecb5](https://github.com/Azure/LogicAppsUX/commit/ce2ecb5f5a7c707c7637c9c984fd8c4265227ca2))
+* **designer:** adding maxWaitingRuns to Concurrency Settings ([#4988](https://github.com/Azure/LogicAppsUX/issues/4988)) ([691fdbb](https://github.com/Azure/LogicAppsUX/commit/691fdbb760bb8978eb9e33e10219e853b1d03490))
+* **Designer:** Tenant selection support for OAuth connection creation ([#4999](https://github.com/Azure/LogicAppsUX/issues/4999)) ([14b7b9c](https://github.com/Azure/LogicAppsUX/commit/14b7b9cb521d1010ceba6988b02ff5b0076010fe))
+* **Templates:** Styling for Header, Tabs, Footer ([#4992](https://github.com/Azure/LogicAppsUX/issues/4992)) ([8a64f12](https://github.com/Azure/LogicAppsUX/commit/8a64f12f40e49aa45d9add0ebf3e1d3ca48da5e2))
+
 ## [4.30.0](https://github.com/Azure/LogicAppsUX/compare/v4.29.0...v4.30.0) (2024-06-17)